Use venv_update to auto update virtualenvs
Use the venv_update tox plugin from tox-pip-extensions
(https://github.com/tox-dev/tox-pip-extensions) to automatically update
tox's Python virtualenvs in development environments whenever
requirements.txt or requirements-dev.in changes.

This fixes an issue that when requirements.txt or requirements-dev.in
changed the virtualenvs in a developer's .tox directory would not be
automatically updated with the new packages. So for example `make dev`
would be running the dev server with a new Python dependency missing, or
a no-longer-used dependency present, or with the old version of an
upgraded dependency. Similarly `make test`'s venv would be out of date,
and the tests would probably be failing.

Developers would have to either delete the `.tox` directory, or run tox
manually with the -r/--recreate option, to fix the situation.

This is a limitation of tox: when a tox.ini file references requirements
files (e.g. `-rrequirements.txt`) tox does not detect changes to those
requirements files and update its virtualenvs. It doesn't look like this
is going to be fixed in tox anytime soon. See this tox issue:

https://github.com/tox-dev/tox/issues/149

A number of solutions are suggested in the above tox issue:

You can use Makefile tricks to trigger updating the venvs but this seems
a bit hacky and requires knowledge of how Makefiles work as well as how
tox.ini files work. I'd rather keep our Makefile as simple as possible
and lean on tox instead.

You can call pip-sync (a command that syncs a virtualenv with a
requirements file) from within your tox.ini file. But this only works
with requirements.txt files produced by pip-compile. Or
requirements-dev.in file is not pip-compiled, nor are out tests-specific
requirements (which we list in tox.ini directly).

Finally, you can use the venv_update tox plugin, an official tox plugin
that was developed to solve this exact problem. This seems to be the
simplest solution.