tox: Crash if tox-pip-extensions not installed
If tox-pip-extensions isn't installed in a developer's environment then
the `tox_pip_extensions_ext_venv_update = true` in tox.ini would just be
silently ignored, tox's virtualenvs would not be automatically updated
when requirements files change, and commands like `make dev` and
`make test` could run with a missing dependency, or with an out of date
version of a dependency. This could be confusing for developers.

It's better to crash, and tell the developer that tox-pip-extensions is
needed, if it isn't installed.