Enable running devserver in Python 3.6 or 3.7
`make dev` runs `tox -e py27-dev` which runs the dev server in Python
2.7. `tox -e py36-dev` will run the dev server in Python 3.6, and
`tox -e py37-dev` will run it in 3.7.

Only Pythons 2.7, 3.6 and 3.7 are possible, and this only works for the
dev server.

The tox.ini code to do this is quite duplicative and boilerplate-y.

There is a much more concise way to write the tox.ini file without all
the duplication, and allowing _any_ of the tox commands to be run with
_any_ version of Python, but it requires the whole tox.ini file to be
rewritten in a different style. This refactoring will be done
sepaarately in a later commit.